remove contact syncer from data layer
