--- core_chat_data/src/androidMain/kotlin/net/thechance/mena/core_chat/data/contacts/ContactSyncException.kt
@@ -1,6 +1,0 @@
-package net.thechance.mena.core_chat.data.contacts
-
-enum class ContactSyncException {
-    PERMISSION_DENIED_EXCEPTION,
-    NETWORK_EXCEPTION
-}
