--- core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/contacts/source/remote/ContactSyncer.kt
@@ -1,5 +1,0 @@
-package net.thechance.mena.core_chat.data.contacts.source.remote
-
-interface ContactSyncer {
-    suspend fun sync()
-}
