+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/di/chatDataModule.kt
@@ -3,5 +3,5 @@ package net.thechance.mena.core_chat.data.di
 import org.koin.dsl.module
 
 val chatDataModule = module {
-    includes(networkModule, repositoryModule, dataProviderModule,contactSyncerModule,platformModule)
+    includes(networkModule, repositoryModule, dataProviderModule)
 }
