--- core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/di/contactSyncerModule.kt
@@ -1,13 +1,0 @@
-package net.thechance.mena.core_chat.data.di
-
-
-import net.thechance.mena.core_chat.data.contacts.source.remote.ContactSyncer
-import org.koin.core.scope.Scope
-import org.koin.dsl.module
-
-
-internal val contactSyncerModule = module {
-    single { createContactSyncerModule() }
-}
-
-expect fun Scope.createContactSyncerModule(): ContactSyncer
