--- core_chat_data/src/iosMain/kotlin/net/thechance/mena/core_chat/data/di/contactSyncerModule.ios.kt
@@ -1,10 +1,0 @@
-package net.thechance.mena.core_chat.data.di
-
-import net.thechance.mena.core_chat.data.contacts.ContactSyncerImpl
-import net.thechance.mena.core_chat.data.contacts.source.remote.ContactSyncer
-import org.koin.core.scope.Scope
-
-actual fun Scope.createContactSyncerModule(): ContactSyncer {
-    return ContactSyncerImpl(get(),get())
-}
-
