+++ core_chat_presentation/src/androidMain/kotlin/net/thechance/mena/core_chat/presentation/di/platformModule.android.kt
@@ -1,0 +1,7 @@
+package net.thechance.mena.core_chat.presentation.di
+
+import net.thechance.mena.core_chat.presentation.sync.ContactSyncWorker
+import org.koin.androidx.workmanager.dsl.workerOf
+import org.koin.dsl.module
+
+actual val platformModule = module { workerOf(::ContactSyncWorker) }
