+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/di/ChatPresentationModule .kt
@@ -3,5 +3,5 @@ package net.thechance.mena.core_chat.presentation.di
 import org.koin.dsl.module
 
 val chatPresentationModule = module {
-    includes(viewModelModule, apiModule, contactSyncerModule)
+    includes(viewModelModule, apiModule, contactSyncerModule, platformModule)
 }
