+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/di/platformModule.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.core_chat.data.di
+package net.thechance.mena.core_chat.presentation.di
 
 import org.koin.core.module.Module
