+++ core_chat_presentation/src/iosMain/kotlin/net/thechance/mena/core_chat/presentation/di/platformModule.ios.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.core_chat.data.di
+package net.thechance.mena.core_chat.presentation.di
 
 import org.koin.dsl.module
